refactor(server): extract /api/analytics/* into a router module
Eighth route-group extraction — the biggest yet: 11 routes scattered
across server.js (2811→14106, 1,395 lines) + the analytics-only
refreshGSCToken (GSC OAuth) helper, into src/server/routes/analytics.js,
mounted `app.use('/api/analytics', analyticsRouter)`.

Auth: MIXED → per-route. 9 routes are requireAuth; the 2 "open" ones
(POST /extract-patterns/:id, POST /sync/:id) do the cron-bypass-OR-Clerk
pattern inline (adminPassword bypass else jwtVerify) — so they keep no
mount middleware but the module imports jwtVerify + clerkJWKS for the
inline check.

Deps — all already extracted: pool, anthropic, safeParseLLM,
requireAuth/verifyBrandAccess/clerkJWKS (auth), callZernio (zernio),
buildXOAuthHeader/refreshXOAuth2Token (x), buildGhostJWT (ghost) — the
sync path pulls analytics from X (token refresh) + Ghost + Zernio. Plus
jwtVerify (jose). refreshGSCToken moved in (analytics-only).

Handler bodies verbatim; only registration lines changed.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (it caught jwtVerify/clerkJWKS/
  refreshXOAuth2Token/buildGhostJWT — added them)
- module boot-load (import()) -> resolves; named exports all present
- route-inventory guard -> 213 byte-identical (11 paths)
- vitest -> 93/93

Rollback: revert; verbatim multi-span move + helper.
